modules folder created, code refactored by creating a module for the function used to populate HTML for each task
