+++ .eslintrc.json
@@ -17,7 +17,8 @@
       "eol-last": "off",
       "import/extensions": [ 1, {
         "js": "always", "json": "always"
-      }]
+      }],
+      "linebreak-style": ["error", "window"]
     },
     "ignorePatterns": [
       "dist/",
